src/backend/python/api-write/src/infrastructure/database.py
import os

from dotenv import load_dotenv
from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
from sqlalchemy.orm import sessionmaker, declarative_base

load_dotenv()

DATABASE_URL = os.getenv("POSTGRES_URL", "postgresql+asyncpg://admin:root@localhost:5432/python")

engine = create_async_engine(DATABASE_URL, echo=True)

AsyncSessionLocal = sessionmaker(
    bind=engine,
    expire_on_commit=False,
    class_=AsyncSession
)

async def get_postgres_session():
    async with AsyncSessionLocal() as session:
        yield session



