src/backend/python/api-write/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
sqlalchemy[asyncio]==2.0.20
asyncpg==0.29.0
motor==3.7.1
pydantic==1.10.12
loguru==0.7.0
aio_pika
pika
python-dotenv==1.0.1
pymongo==4.9.1
pytest==7.4.2
pytest-asyncio==0.21.0
